% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionOccurrenceId_OccurrenceTable.R
\name{conditionOccurrenceId}
\alias{conditionOccurrenceId}
\title{'conditionOccurrenceId'}
\usage{
conditionOccurrenceId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating conditionOccurrenceId
}
\description{
conditionOccurrenceId function indicates conditionOccurrenceId which needs to be joined with OMOP-CDM
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
conditionOccurrenceId(DICOMList)
}
