% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiologyDB.R
\name{createRadiologyImage}
\alias{createRadiologyImage}
\title{createRadiologyImage}
\usage{
createRadiologyImage(data)
}
\arguments{
\item{data}{Data frame containing image information}
}
\description{
This function creates the Image table based on Radiology CDM.
You need to create an Occurrence table based on Radiology CDM in your DBMS as a preliminary work,
and if there is no table in advance, it will output an error.
}
\examples{
############################# Example Code: Create Radiology from RDS.. #############################

dirpath <- "/home/ohdsi/Documents/DICOM-header"

files <- list.files(path = dirpath, full.names = TRUE, recursive = TRUE, pattern = "\\\\.rds$")
tpb <- txtProgressBar(min = 0, max = length(files), style = 3)

for(i in 1:length(files)) {
  setTxtProgressBar(tpb, i)
  df <- createRadiologyImage(data = readRDS(files[i]))
  View(df)
}

######################################## Example Code: END ##########################################
}
\seealso{
Radiology CDM Wiki: https://github.com/NEONKID/DicomHeaderExtractionModule/wiki
}
\author{
Neon K.I.D
}
