% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiologyDB.R
\name{createRadiologyOccurrence}
\alias{createRadiologyOccurrence}
\title{createRadiologyOccurrence}
\usage{
createRadiologyOccurrence(path)
}
\arguments{
\item{path}{Enter the folder of radiation data to be extracted.}
}
\description{
This function creates the Occurrence table based on Radiology CDM.
You need to create an Occurrence table based on Radiology CDM in your DBMS as a preliminary work,
and if there is no table in advance, it will output an error.
}
\examples{
############################# Example Code: Create Radiology from RDS.. #############################

dirpath <- "/home/ohdsi/Documents/DICOM-header"

dirs <- list.dirs(path = dirpath, full.names = TRUE, recursive = FALSE)
tpb <- txtProgressBar(min = 0, max = length(dirs), style = 3)

for(i in 1:length(dirs)) {
  files <- list.files(path = dirs[i], pattern = "\\\\.rds$", recursive = TRUE)
  if(length(files) == 0) next
  setTxtProgressBar(tpb, i)
  df <- createRadiologyOccurrence(dirs[i])
  View(df)
}

######################################## Example Code: END ##########################################
}
\seealso{
Radiology CDM Wiki: https://github.com/NEONKID/DicomHeaderExtractionModule/wiki
}
\author{
Neon K.I.D
}
