% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICOMFileModule.R
\docType{data}
\name{DcmFileModule}
\alias{DcmFileModule}
\title{DcmFileModule Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DcmFileModule$new(path, savePathRoot)
}
\arguments{
\item{path}{Path with DICOM image file}

\item{savePathRoot}{Top path to store extracted metadata}
}
\description{
This class creates a dataset by compressing a batch of DICOM files into a single RDS file.
Each DICOM file is so large that it takes a lot of time to read it, and it is a function that extracts only its metadata.
}
\examples{
############################# Example Code: Create DICOM RDS File ###################################

# Path is directory folder
path <- "/home/ohdsi/Pictures/Dicom-Images/DCM-12411"

# savePathRoot is directory folder
savePathRoot <- "/home/ohdsi/DICOM-header"

DFM <- DcmFileModule$new(path = path, savePathRoot = savePathRoot)

# rootPathCount is a parameter that determines which of the current DICOM file paths to cut.
# If you do not want to cut this parameter, you can omit this parameter.
DFM$dcmToRDS(rootPathCount = 4, verbose = FALSE)

######################################## Example Code: END ##########################################
}
\author{
Neon K.I.D
}
\keyword{datasets}
