% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DicomPath_ImageTable.R
\name{DicomPath}
\alias{DicomPath}
\title{'DicomPath'}
\usage{
DicomPath(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list indicating path of each DICOM files
}
\description{
DicomPath function indicates path of each DICOM files
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
lapply(DICOMList)
}
