% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DicomRDS.R
\docType{data}
\name{DicomRDS}
\alias{DicomRDS}
\title{DicomRDS Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DicomRDS$new(data)
}
\arguments{
\item{data}{Data frame imported from DICOM RDS file}
}
\description{
This class is a class that imports various metadata by reading an RDS file containing DICOM information.
There are basically functions that import data related to Radiology CDM.
}
\examples{
############################# Example Code: Read DICOM RDS File   ###################################

rdsFile <- "/home/ohdsi/DICOM-header/header.rds"

data <- readRDS(file = rdsFile)

# If single data (only one metadata in dicom file )
dcmRDS <- DicomRDS$new(data)

# Create Occurrence ID for radiology image
roID <- dcmRDS$createOccurrenceID()

# Get PatientID
patientid <- dcmRDS$getPatientID()

# This radiology contast or non-contrast
isContrast <- dcmRDS$isPost4BrainCT()

# If Multi data
for(i in 1:length(data)) {
  dcmRDS <- DicomRDS$new(data[[i]])
  roID <- dcmRDS$createOccurrenceID()
  patientid <- dcmRDS$getPatientID()
  isContrast <- dcmRDS$isPost4BrainCT()
  df <- data.frame(roID, patientid, isContrast)
}

View(df)

######################################## Example Code: END ##########################################
}
\seealso{
https://github.com/NEONKID/RCDM-ETL/wiki
}
\author{
Neon K.I.D
}
\keyword{datasets}
