% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageId_ImageTable.R
\name{imageId}
\alias{imageId}
\title{'imageId'}
\usage{
imageId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing anonymized SOPInstanceUID of DICOM
}
\description{
imageId function will read SOPInstanceUID metadata of all DICOM files read by DICOMHeaderList function
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageId(DICOMList)
}
