% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageType_ImageTable.R
\name{imageType}
\alias{imageType}
\title{'imageType'}
\usage{
imageType(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing type of each DICOM image
}
\description{
imageType function will represent type of each image (ex. Primary or Secondary)
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
imageType(DICOMList)
}
