% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personId_ImageTable.R
\name{personId}
\alias{personId}
\title{'personId'}
\usage{
personId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing anonymized PatientID of DICOM
}
\description{
personId function will read PatientID metadata of all DICOM files read by DICOMHeaderList function
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
personId(DICOMList)
}
