% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiologyDB.R
\docType{data}
\name{RadDB}
\alias{RadDB}
\title{RadDB Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RadDB$new(core)
}
\arguments{
\item{core}{Number of cores to use}
}
\description{
This is a class that creates an RCDM using an extracted RDS file.
Parallel processing is supported only in the current Occurrence table,
and it is recommended that the image table not be processed in parallel
because of the computer science problem that the number of images in the image table does not match.
Even if you set Pararell through the constructor, it is not supported by the Image method.
}
\examples{
############################# Example Code: Convert Radiology CDM  ###################################

# Create Radiology_Occurrence table for rds Path
rdsPath <- "/home/ohdsi/Radiology/rdsfiles"

# Create RadDB object,,
# If using pararell, require pararell package
RDB <- RadDB$new(core = parallel::detectCores() - 1)

# Get Radiology_Occurrence table
occur <- RDB$createRadiologyOccurrence(path = rdsPath)

# Create Radiology_Image table for read RDS file
rdsFile <- "/home/ohdsi/DICOM-header/header.rds"
data <- readRDS(file = rdsFile)

# Get Radiology_Image table
img <- RDB$createRadiologyImage(data = data)

######################################## Example Code: END ##########################################
}
\seealso{
https://github.com/OHDSI/Radiology-CDM/wiki
}
\author{
Neon K.I.D
}
\keyword{datasets}
