% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationExposureTime_OccurrenceTable.R
\name{radiationExposureTime}
\alias{radiationExposureTime}
\title{'radiationExposureTime'}
\usage{
radiationExposureTime(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list indicating duration of radiation exposed by each shoot
}
\description{
radiationExposureTime function indicates duration of radiation exposed by each shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiationExposureTime(DICOMList)
}
