% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyDirpath_OccurrenceTable.R
\name{radiologyDirpath}
\alias{radiologyDirpath}
\title{'radiologyDirpath'}
\usage{
radiologyDirpath(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating path which includes a single occurrence
}
\description{
radiologyDirpath function indicates path which includes a single occurrence
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyDirpath(DICOMList)
}
