% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyImageTable.R
\name{radiologyImageTable}
\alias{radiologyImageTable}
\title{'radiologyImageTable'}
\usage{
radiologyImageTable(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
radiologyImageTable in a dataframe format
}
\description{
radiologyImageTable function shows you radiologyImageTable in a dataframe format
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyImageTable(DICOMList)
}
