% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyOccurrenceId2_OccurrenceTable.R
\name{radiologyOccurrenceId2}
\alias{radiologyOccurrenceId2}
\title{'radiologyOccurrenceId2'}
\usage{
radiologyOccurrenceId2(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating occurrence ID of each shoot
}
\description{
radiologyOccurrenceId2 function indicates occurrence ID of each shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyOccurrenceId2(DICOMList)
}
