% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyPhaseConceptId_ImageTable.R
\name{radiologyPhaseConceptId}
\alias{radiologyPhaseConceptId}
\title{'radiologyPhaseConceptId'}
\usage{
radiologyPhaseConceptId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list containing phase of each DICOM image
}
\description{
radiologyPhaseConceptId function will represent phase of each image
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyPhaseConceptId(DICOMList)
}
