% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiologyProtocolConceptId_OccurrenceTable.R
\name{radiologyProtocolConceptId}
\alias{radiologyProtocolConceptId}
\title{'radiologyProtocolConceptId'}
\usage{
radiologyProtocolConceptId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating path protocol of each shoot
}
\description{
radiologyProtocolConceptId function indicates protocol of each shoot
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
radiologyProtocolConceptId(DICOMList)
}
