% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitOccurrenceId_OccurrenceTable.R
\name{visitOccurrenceId}
\alias{visitOccurrenceId}
\title{'visitOccurrenceId'}
\usage{
visitOccurrenceId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A list indicating visitOccurrenceId
}
\description{
visitOccurrenceId function indicates visitOccurrenceId which needs to be joined with OMOP-CDM
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
visitOccurrenceId(DICOMList)
}
