% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonId_OccurrenceTable.R
\name{PersonId}
\alias{PersonId}
\title{'PersonId'}
\usage{
PersonId(DICOMList)
}
\arguments{
\item{DICOMList}{you can put it like this and then run the function : DICOMList<-DICOMHeaderList(DICOMFolderPath)}
}
\value{
A dataframe indicating patient ID and Occurrence ID of each occurrences
}
\description{
PersonId function indicates patient ID of each occurrences
}
\examples{
DICOMList<-DICOMHeaderList(DICOMFolderPath)
PersonId(DICOMList)
}
